pyproject.toml
[tool.poetry]
name = "manager"
version = "0.1.0"
description = ""
authors = ["Tamagil <tamagiltamagil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "==5.0.*"
djangorestframework = "3.15"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
